+++ libft/quat/quat_normalize.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   normalize_quat.c                                   :+:      :+:    :+:   */
+/*   quat_normalize.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: lsohler <lsohler@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/13 18:04:54 by lsohler           #+#    #+#             */
-/*   Updated: 2023/09/13 18:05:54 by lsohler          ###   ########.fr       */
+/*   Updated: 2023/10/27 15:38:36 by lsohler          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 t_quat	quat_normalize(t_quat q)
 {
-	double	magnitude;
+	float	magnitude;
 
 	magnitude = sqrt(q.w * q.w + q.x * q.x + q.y * q.y + q.z * q.z);
 	q.w /= magnitude;
